# planning-service/openapi.json
{
  "components": {
    "schemas": {
...
      "AirportSearch": {
        "properties": {
          "airports": {
            "items": {
              "$ref": "#/components/schemas/AirportSuggestion"
            },
            "title": "Airports",
            "type": "array"
          }
        },
        "required": [
          "airports"
        ],
        "title": "AirportSearch",
        "type": "object"
      },
...
      "AirportsInView": {
        "properties": {
          "airports": {
            "items": {
              "$ref": "#/components/schemas/AirportPin"
            },
            "title": "Airports",
            "type": "array"
          }
        },
        "required": [
          "airports"
        ],
        "title": "AirportsInView",
        "type": "object"
      },
...
      "CheckpointNoteRequest": {
        "properties": {
          "departure_ident": {
            "title": "Departure Ident",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "destination_ident": {
            "title": "Destination Ident",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "stops": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Stops",
            "type": "array"
          }
        },
        "required": [
          "departure_ident",
          "destination_ident",
          "lat",
          "lon",
          "description"
        ],
        "title": "CheckpointNoteRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NearestAirports": {
        "properties": {
          "airports": {
            "items": {
              "$ref": "#/components/schemas/NearestAirport"
            },
            "title": "Airports",
            "type": "array"
          }
        },
        "required": [
          "airports"
        ],
        "title": "NearestAirports",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Wingtip Maps planner",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/airport/{ident}": {
      "get": {
        "description": "One US airport's card, by any ident it goes by (C81, KC81, KDLH).\n\n`airspace_class` is the class of the airspace at the surface at the\nfield: B, C or D, E in a Class E surface area, G where the ground is\nuncontrolled (vfr.airspace.surface_class_at).\n`towered` is whether it lists a tower frequency. A field with no\nreporting station has no `metar`; `weather_unavailable` is set only\nwhen the weather service could not be asked.",
        "operationId": "airport_place_api_airport__ident__get",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirportPlace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Airport Place"
      }
    },
    "/api/airports/in-view": {
      "get": {
        "description": "The landing fields inside the map's view, the biggest first --\nat most `limit` of them, since a whole state holds thousands and the\nmap only lays a tap target over each -- with each one's flight\ncategory from its METAR, for a weather chip on the ones that report.\nThe METARs are the national cache already in memory; with the\nweather service out the fields come back without, as fields with no\nstation do. With `reporting`, only the fields that report, before\nthe limit: what the map asks once for a whole route's box, so its\nchips are there before the pilot zooms in on any of it.",
        "operationId": "airports_in_view_api_airports_in_view_get",
        "parameters": [
          {
            "in": "query",
            "name": "south",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "South",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "west",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "West",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "north",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "North",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "east",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "East",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "reporting",
            "required": false,
            "schema": {
              "default": false,
              "title": "Reporting",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirportsInView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Airports In View"
      }
    },
    "/api/airports/nearest": {
      "get": {
        "description": "The landing fields nearest a position -- own ship's, for the map's\nNearest -- the nearest first: how far and which way, each one's\nlongest open runway and its flight category where it reports.",
        "operationId": "nearest_airports_api_airports_nearest_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Lon",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearestAirports"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nearest Airports"
      }
    },
    "/api/airports/search": {
      "get": {
        "description": "DEP/DEST's own autocomplete -- every airport whose ident or name\nstarts with `q`, for the route inputs to suggest as a pilot types.\nRuns against the same in-memory OurAirports table the real lookup\nuses, not a second data source that could drift from it. With\n`fixes`, a stop's: the named fixes whose ident starts with it too,\nafter the airports -- VFR waypoints first (vfr.fixes).",
        "operationId": "airport_search_api_airports_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fixes",
            "required": false,
            "schema": {
              "default": false,
              "title": "Fixes",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirportSearch"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Airport Search"
      }
    },
    "/api/briefing": {
      "get": {
        "description": "Everything the nav log's own leg math doesn't cover: adverse\nconditions (SIGMET/AIRMET), current conditions (METAR) and\nforecast (TAF-derived ceiling/visibility) along the route, and\neach airport's runways and radio frequencies.\n\nThe forecast is for the flight: from `depart` (now when not given;\nUTC when naive) to an hour past arrival, `ete_min` after it (two\nhours when not given). It used to be read at the moment of asking,\nwhatever time the pilot was planning to go. With stops (`stops`,\n\"KDSM,KLNK\"), along every hop and for every airport landed at.",
        "operationId": "briefing_api_briefing_get",
        "parameters": [
          {
            "in": "query",
            "name": "dep",
            "required": true,
            "schema": {
              "title": "Dep",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dest",
            "required": true,
            "schema": {
              "title": "Dest",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stops",
            "required": false,
            "schema": {
              "default": "",
              "title": "Stops",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depart",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Depart"
            }
          },
          {
            "in": "query",
            "name": "ete_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ete Min"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Briefing"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Briefing"
      }
    },
...
    "/api/chart-tile/{kind}/{z}/{x}/{y}.png": {
      "get": {
        "description": "Any kind of chart by its key (`chart_layers` on the course lists\nthem): `sec`, `tac`, `ifr_low`, `ifr_high`, `ifr_area`. What the\nmap's layers draw from.\n\nThe FAA's own GeoTIFF of each sheet, rendered into a {z}/{x}/{y} tile\npyramid (vfr.charts -- downloaded once per 56-day chart cycle) and\ncached on disk; the same tiles the detector reads. A tile pyramid is\nwhat makes each chart a plain Leaflet tile layer: edge-only fetches\non a pan, the previous zoom's tiles scaled under the zoom animation,\na prefetch ring. 404 where no sheet covers the tile, which the layer\nleaves blank. The sectional and the TAC had routes of their own\nunder older names, which nothing called any more.",
        "operationId": "chart_tile_api_chart_tile__kind___z___x___y__png_get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "z",
            "required": true,
            "schema": {
              "title": "Z",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "x",
            "required": true,
            "schema": {
              "title": "X",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "y",
            "required": true,
            "schema": {
              "title": "Y",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chart Tile"
      }
    },
...
    "/api/checkpoint-notes": {
      "post": {
        "description": "A pilot's own edit to a checkpoint's identification note: what\nthey see at that place from now on. Other pilots keep the shared\nnote; where nobody can sign in, this becomes the shared note.",
        "operationId": "save_checkpoint_note_api_checkpoint_notes_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Pilot-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pilot-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointNoteSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Checkpoint Note"
      }
    },
    "/api/checkpoint-notes/generate": {
      "post": {
        "description": "One \"how to spot it\" line per checkpoint, as newline-delimited\nJSON (each line one app.schemas.CheckpointNoteMessage), so a slow\nLLM call on checkpoint 3 does not hold up checkpoints 1 and 2 that\nalready arrived.\n\nTwo different kinds of failure, reported two different ways. A\nsingle checkpoint's own LLM call failing for a reason specific to\nit does not take the rest of the stream down -- every other\ncheckpoint is independent and still worth generating, so that one\ngets its own per-checkpoint \"error\" line and the loop moves on.\nGLOBAL_ANTHROPIC_ERRORS is the opposite case: a bad key or an\nexhausted rate limit fails identically for every checkpoint, so\nit's reported once as a single stream-level \"error\" (for one\nbanner instead of the same text repeated per row) -- but the\nstream itself keeps going: every remaining checkpoint still gets\nits own per-checkpoint \"error\" line too, just without spending a\ncall to learn what's already known. A pilot still gets a row to\ntype a note into for every checkpoint, not just the ones that\nhappened to come before the failure.\n\nA POST, not a GET: every checkpoint without a note is a billed\nClaude call and a write, which a GET invited from anything that\nprefetches a link, and which the gateway's rules left public in\nevery deployment. The caller's own edit is what they see where they\nmade one; generation only ever fills the shared note.",
        "operationId": "describe_checkpoints_api_checkpoint_notes_generate_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Pilot-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pilot-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateNotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Describe Checkpoints"
      }
    },
    "/api/checkpoints": {
      "get": {
        "description": "Scored candidates and the subset worth flying, hop by hop along\nthe whole route: fast once each hop's chart has been read, a few\nseconds for its first time.",
        "operationId": "checkpoints_api_checkpoints_get",
        "parameters": [
          {
            "in": "query",
            "name": "dep",
            "required": true,
            "schema": {
              "title": "Dep",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dest",
            "required": true,
            "schema": {
              "title": "Dest",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stops",
            "required": false,
            "schema": {
              "default": "",
              "title": "Stops",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Checkpoints"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Checkpoints"
      }
    },
...
    "/api/classify": {
      "get": {
        "description": "What the chart draws at a point, so a hand-marked checkpoint is\ncategorised from the pixels rather than from whatever the dropdown\nhappened to be left on.",
        "operationId": "classify_api_classify_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "title": "Lon",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classify"
      }
    },
    "/api/course": {
      "get": {
        "description": "Just the course line and its endpoints, through any stops.\n\nSeparate from detection so the chart can draw a line the instant two\nidents are entered. Reading tiles takes seconds even on the fast\npath, and there is no reason a pilot should watch an empty map for\nthem.",
        "operationId": "course_api_course_get",
        "parameters": [
          {
            "in": "query",
            "name": "dep",
            "required": true,
            "schema": {
              "title": "Dep",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dest",
            "required": true,
            "schema": {
              "title": "Dest",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stops",
            "required": false,
            "schema": {
              "default": "",
              "title": "Stops",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Course"
      }
    },
...
  }
}